fix2

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: tishihar <tishihar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 23:03:30 by tishihar          #+#    #+#             */
-/*   Updated: 2024/05/18 15:59:13 by tishihar         ###   ########.fr       */
+/*   Updated: 2024/06/09 13:14:29 by tishihar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (*(s++))
